folder reorganization and gitignore updates
