+++ .gitignore
@@ -26,6 +26,8 @@ share/python-wheels/
 *.egg
 MANIFEST
 _version.py
+conda.recipe/*
+!conda.recipe/meta.yaml
 # PyInstaller
 #  Usually these files are written by a python script from a template
 #  before PyInstaller builds the exe, so as to inject date/other infos into it.
