+++ src/inspired/gui/help_dft.ui
