+++ src/inspired/gui/inspired.ui
