+++ src/inspired/gui/latent_space_model.py
