+++ src/inspired/gui/oclimax.ui
