+++ src/inspired/gui/oclimax_params.py
