+++ src/inspired/gui/pandas_model.py
