+++ src/inspired/gui/set_paths.py
@@ -1,5 +1,5 @@
 from qtpy.QtWidgets import (QFileDialog, QDialog)
-from inspired.ui_set_paths import Ui_SetPaths
+from inspired.gui.ui_set_paths import Ui_SetPaths
 import os
 import sys
