+++ src/inspired/gui/set_paths.ui
