+++ src/inspired/gui/ui_help_dft.py
