+++ src/inspired/gui/ui_oclimax.py
