+++ src/inspired/gui/ui_set_paths.py
